requirements.txt
Django==3.2.3
psycopg2-binary==2.8.6
pytz==2021.1
python-decouple==3.4
gunicorn==20.1.0
asgiref==3.3.4
sqlparse==0.4.1
whitenoise==5.2.0